refactor(@crates/eur-activity): consolidate the strategy creation to a single factory file
